Update Laurel's Program 2
not sure if my statement "input = kbd.readLine()" is correct for reading the input/output files. I have been piece by piece working on the tokens, so just comment that out if you need to run it bc there are a few errors in it right now. One thing if you could look at it, is line 109, the character.isdigit(token1). is that testing a single character or a string of characters? let me know what you think so far and if you can get rid of any of those errors. I will work on it some more later today. thanks!