Update G3E1.java
I edited the format, made it a little bit nicer looking. I added in the totals at the bottom for output to the screen and to the file. I think that the average for each interval is wrong. I think we just need to have the 1st number / 2nd number. So, the mileage divided by the fuel consumed. Looking at my input data file, we would have 62358/.125 = 498864. We can talk about it tomorrow. Also, we need to make sure the user cannot overwrite our input file, Make sure they don't enter the same file for each input and output.