Update Program3.java
Added in checkcopy function in order to copy files when appropriate. Also added brainstorming comments for the "+" for the directory.  We might still need to add a stop function at the end to close the windows.  I think the target file is messing up now I'm not sure.  It still compiles. Hope some of it helps.